Refs #3008 AlgorithmsTest builds.

+++ Code/Mantid/Framework/Algorithms/test/BinaryOperationTest.h
@@ -122,7 +122,7 @@ public:
   void testMaskedSpectraPropagation()
   {
     const int nHist = 10,nBins=20;
-    std::set<int> masking;
+    std::set<int64_t> masking;
     masking.insert(0);
     masking.insert(2);
     masking.insert(4);
